BE-#49 refactor: 리팩토링 도구를 이용해서 exception 패키지와 error 패키지 분리
